Break `h.auth` package into pieces
Moves the authentication policy into `h.auth.policy` and a bunch of
helpers into `h.auth.util`.

The only externally-used piece of code is the `effective_principals`
helper function which is exported at the top level of the package.
