+++ h/auth/test/util_test.py
@@ -5,7 +5,7 @@ import mock
 from pyramid import testing
 from pyramid import security
 
-from h import auth
+from h.auth import util
 from h.api.models.token import API_TOKEN_PREFIX
 
 
@@ -20,7 +20,7 @@ effective_principals_fixtures = pytest.mark.usefixtures('accounts', 'group_princ
 def test_effective_principals_when_user_is_None(accounts):
     accounts.get_user.return_value = None
 
-    principals = auth.effective_principals('acct:jiji@hypothes.is',
+    principals = util.effective_principals('acct:jiji@hypothes.is',
                                            testing.DummyRequest())
 
     assert principals == [security.Everyone]
@@ -33,7 +33,7 @@ def test_effective_principals_returns_Authenticated(accounts, group_principals):
     accounts.get_user.return_value = mock.Mock(admin=False, staff=False)
     group_principals.return_value = []
 
-    principals = auth.effective_principals('acct:jiji@hypothes.is',
+    principals = util.effective_principals('acct:jiji@hypothes.is',
                                            testing.DummyRequest())
 
     assert security.Authenticated in principals
@@ -46,7 +46,7 @@ def test_effective_principals_returns_userid(accounts, group_principals):
     accounts.get_user.return_value = mock.Mock(admin=False, staff=False)
     group_principals.return_value = []
 
-    principals = auth.effective_principals('acct:jiji@hypothes.is',
+    principals = util.effective_principals('acct:jiji@hypothes.is',
                                            testing.DummyRequest())
 
     assert 'acct:jiji@hypothes.is' in principals
@@ -57,7 +57,7 @@ def test_effective_principals_when_user_is_admin(accounts, group_principals):
     accounts.get_user.return_value = mock.Mock(admin=True, staff=False)
     group_principals.return_value = []
 
-    principals = auth.effective_principals('acct:jiji@hypothes.is',
+    principals = util.effective_principals('acct:jiji@hypothes.is',
                                            testing.DummyRequest())
 
     assert 'group:__admin__' in principals
@@ -68,7 +68,7 @@ def test_effective_principals_when_user_is_staff(accounts, group_principals):
     accounts.get_user.return_value = mock.Mock(admin=False, staff=True)
     group_principals.return_value = []
 
-    principals = auth.effective_principals('acct:jiji@hypothes.is',
+    principals = util.effective_principals('acct:jiji@hypothes.is',
                                            testing.DummyRequest())
 
     assert 'group:__staff__' in principals
@@ -79,7 +79,7 @@ def test_effective_principals_when_user_has_groups(accounts, group_principals):
     accounts.get_user.return_value = mock.Mock(admin=False, staff=False)
     group_principals.return_value = ['group:abc123', 'group:def456']
 
-    principals = auth.effective_principals('acct:jiji@hypothes.is',
+    principals = util.effective_principals('acct:jiji@hypothes.is',
                                            testing.DummyRequest())
 
     for group in group_principals.return_value:
@@ -91,7 +91,7 @@ def test_effective_principals_with_staff_admin_and_groups(accounts, group_princi
     accounts.get_user.return_value = mock.Mock(admin=True, staff=True)
     group_principals.return_value = ['group:abc123', 'group:def456']
 
-    principals = auth.effective_principals('acct:jiji@hypothes.is',
+    principals = util.effective_principals('acct:jiji@hypothes.is',
                                            testing.DummyRequest())
 
     for principal in [security.Everyone,
@@ -111,13 +111,13 @@ def _mock_group(pubid):
 def test_group_principals_with_no_groups():
     user = mock.Mock(groups=[])
 
-    assert auth.group_principals(user) == []
+    assert util.group_principals(user) == []
 
 
 def test_group_principals_with_one_group():
     user = mock.Mock(groups=[_mock_group('pubid1')])
 
-    assert auth.group_principals(user) == ['group:pubid1']
+    assert util.group_principals(user) == ['group:pubid1']
 
 
 def test_group_principals_with_three_groups():
@@ -127,7 +127,7 @@ def test_group_principals_with_three_groups():
         _mock_group('pubid3'),
     ])
 
-    assert auth.group_principals(user) == [
+    assert util.group_principals(user) == [
         'group:pubid1',
         'group:pubid2',
         'group:pubid3',
@@ -139,13 +139,13 @@ def test_bearer_token_returns_token():
         'Authorization': 'Bearer ' + API_TOKEN_PREFIX + 'abc123'
     })
 
-    assert auth.bearer_token(request) == API_TOKEN_PREFIX + 'abc123'
+    assert util.bearer_token(request) == API_TOKEN_PREFIX + 'abc123'
 
 
 def test_bearer_token_when_no_Authorization_header():
     request = testing.DummyRequest(headers={})
 
-    assert auth.bearer_token(request) == ''
+    assert util.bearer_token(request) == ''
 
 
 def test_bearer_token_when_Authorization_header_does_not_contain_bearer():
@@ -153,18 +153,18 @@ def test_bearer_token_when_Authorization_header_does_not_contain_bearer():
         'Authorization': API_TOKEN_PREFIX + 'abc123'  # No "Bearer " prefix.
     })
 
-    assert auth.bearer_token(request) == ''
+    assert util.bearer_token(request) == ''
 
 
 @pytest.fixture
 def accounts(request):
-    patcher = mock.patch('h.auth.accounts', autospec=True)
+    patcher = mock.patch('h.auth.util.accounts', autospec=True)
     request.addfinalizer(patcher.stop)
     return patcher.start()
 
 
 @pytest.fixture
 def group_principals(request):
-    patcher = mock.patch('h.auth.group_principals', autospec=True)
+    patcher = mock.patch('h.auth.util.group_principals', autospec=True)
     request.addfinalizer(patcher.stop)
     return patcher.start()
